Update chapter four figures and exercises (#25)
* fix warnings; update chapter four
* add exercise 4.3
* add image for 4.7

+++ harness/src/test/scala/com/robertjneal/rl/AgentTest.scala
@@ -38,24 +38,40 @@ class AgentTest {
 
     assertEquals(agent.e.possibleStateActions, result1.e.possibleStateActions)
     assertEquals(
-      1,
-      result1.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      1L,
+      result1
+        .actionSteps(OneState)
+        .values
+        .map(_.toInt)
+        .sum - actionRewards.size.toLong
     )
     assertEquals(
-      result1.step,
-      result1.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      result1.step.toInt.toLong,
+      result1
+        .actionSteps(OneState)
+        .values
+        .map(_.toInt)
+        .sum - actionRewards.size.toLong
     )
 
     val result2: TabularAgent[Reward] = result1.act
 
     assertEquals(result1.e.possibleStateActions, result2.e.possibleStateActions)
     assertEquals(
-      2,
-      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      2L,
+      result2
+        .actionSteps(OneState)
+        .values
+        .map(_.toInt)
+        .sum - actionRewards.size.toLong
     )
     assertEquals(
-      result2.step,
-      result2.actionSteps(OneState).values.map(_.toInt).sum - actionRewards.size
+      result2.step.toInt.toLong,
+      result2
+        .actionSteps(OneState)
+        .values
+        .map(_.toInt)
+        .sum - actionRewards.size.toLong
     )
   }
 }
